LIR_Django_App/requirements.txt
aiobotocore==2.5.0
aiohttp==3.8.4
aioitertools==0.11.0
aiosignal==1.3.1
asgiref==3.7.2
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
aws-secretsmanager-caching==1.1.1.5
backcall==0.2.0
boto3
botocore
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
decorator==5.1.1
Django==4.2.2
dnspython==2.3.0
executing==1.2.0
filelock==3.12.2
frozenlist==1.3.3
fsspec==2023.6.0
greenlet==2.0.2
huggingface-hub==0.15.1
idna==3.4
ipython==8.14.0
ipython-genutils==0.2.0
ipython-sql==0.5.0
jedi==0.18.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
loguru==0.7.0
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mpmath==1.3.0
multidict==6.0.4
networkx==3.1
nltk==3.8.1
numpy==1.24.3
packaging==23.1
pandas==2.0.2
parso==0.8.3
pickleshare==0.7.5
Pillow==9.5.0
pinecone-client==2.2.2
prettytable==3.8.0
prompt-toolkit==3.0.38
psycopg2
psycopg2-binary
pure-eval==0.2.2
Pygments==2.15.1
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
regex==2023.6.3
requests==2.31.0
s3fs==2023.6.0
s3transfer==0.6.1
safetensors==0.3.1
scikit-learn==1.2.2
scipy==1.10.1
sentence-transformers==2.2.2
sentencepiece==0.1.99
six==1.16.0
SQLAlchemy==2.0.16
sqlparse==0.4.4
stack-data==0.6.2
sympy==1.12
threadpoolctl==3.1.0
tokenizer==3.4.2
tokenizers==0.13.3
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0
traitlets==5.9.0
transformers==4.30.2
typing_extensions==4.6.3
tzdata==2023.3
urllib3==1.26.16
wcwidth==0.2.6
whitenoise==6.4.0
win32-setctime==1.1.0
wrapt==1.15.0
yarl==1.9.2
